we have identified that the issue is related to how the information is going back to the JSON file
