+++ index.html
@@ -37,7 +37,9 @@
             <div class="catForm">
               <label for="catVote">
                 <h1 class="text cardHeader">Team Cat</h1>
-                <div id="cat"></div>
+                <div id="cat">
+                </div>
+                <button type="button" id="catButton">Cats!!!</button>
                 <div id="catPics"></div>
               </label>
             </div>
@@ -48,7 +50,9 @@
             <div class="dogForm">
               <label for="dogVote">
                 <h1 class="text cardHeader">Team Dog</h1>
-                <div id="dog"></div>
+                <div id="dog">
+                </div>
+                <button  type="button" id="dogButton">Dogs!!!</button>
                 <div id="dogPics"></div>
               </label>
             </div>
